Implemented rec_reveal to test environment and ported to main program
It's working! A couple of issues:

The current implementation of rec_reveal in Gameboard does not work with
the currently commented-out reveal() function in Tiles. We may not need it
in Tiles.

We weren't properly keeping track of what has and hasn't been revealed
in the recursive function.  I believe the intention was to have that be
taken care of in the Tiles' own reveal() function, but I believe it's
easier to just let the Gameboard rec_reveal() function handle it.  If
anyone disagrees I am happy to split it back up into two functions.